✨ Feat: 즐겨찾기에서 다이어리로 음식 등록 api 연결 (#37)

+++ app/src/main/java/com/example/snack4diet/api/createFoodFromBookmark/CreateFoodFromBookmark.kt
@@ -1,6 +1,6 @@
 package com.example.snack4diet.api.createFoodFromBookmark
 
 data class CreateFoodFromBookmark(
-    val favoriteFoodId: Int,
-    val userId: Int
+    val favoriteFoodId: Long,
+    val userId: Long
 )
